Endrer til Int og Double mot frontend Mer robust skalBetaleTrygdeavgiftTilNav

+++ frontend-api/src/main/kotlin/no/nav/melosys/tjenester/gui/dto/trygdeavgift/TrygdeavgiftsperioderDto.kt
@@ -2,17 +2,21 @@ package no.nav.melosys.tjenester.gui.dto.trygdeavgift
 
 import no.nav.melosys.domain.avgift.Trygdeavgiftsperiode
 import no.nav.melosys.domain.kodeverk.Trygdedekninger
-import java.math.BigDecimal
-import java.math.BigInteger
 import java.time.LocalDate
 
 data class TrygdeavgiftsperioderDto(
     val fom: LocalDate,
     val tom: LocalDate,
     val trygdedekning: Trygdedekninger,
-    val avgiftssats: BigDecimal,
-    val avgiftPerMd: BigInteger
+    val avgiftssats: Double,
+    val avgiftPerMd: Int
 ) {
     constructor(trygdeavgiftsperiode: Trygdeavgiftsperiode) :
-        this(trygdeavgiftsperiode.periodeFra, trygdeavgiftsperiode.periodeTil, trygdeavgiftsperiode.hentGjeldendeTrygdedekning(), trygdeavgiftsperiode.trygdesats, trygdeavgiftsperiode.trygdeavgiftsbeløpMd)
+        this(
+            trygdeavgiftsperiode.periodeFra,
+            trygdeavgiftsperiode.periodeTil,
+            trygdeavgiftsperiode.hentGjeldendeTrygdedekning(),
+            trygdeavgiftsperiode.trygdesats.toDouble(),
+            trygdeavgiftsperiode.trygdeavgiftsbeløpMd.toInt()
+        )
 }
